finished token validation

+++ controllers/user.controller.js
@@ -1,4 +1,5 @@
 const UserService = require("../services/user.service");
+const TokenService = require("../services/token.service");
 const UserDto = require("../dto/user.dto");
 const ApiError = require("../exceptions/api.error");
 
@@ -35,14 +36,22 @@ exports.getUser = async (req, res, next) => {
 exports.deleteUser = async (req, res, next) => {
   const { id } = req.params;
   try {
+    if (req.user.id !== id) {
+      throw ApiError.Forbidden("No privileges to delete this user");
+    }
     const user = await UserService.deleteUser(id);
     res.status(200).json({ ...new UserDto(user) });
-  } catch (error) {}
+  } catch (error) {
+    next(error);
+  }
 };
 
 exports.updateUser = async (req, res, next) => {
   const { id } = req.params;
   try {
+    if (req.user.id !== id) {
+      throw ApiError.Forbidden("No privileges to update this user");
+    }
     const user = await UserService.updateUser(id, req.body);
     res.status(200).json({ ...new UserDto(user) });
   } catch (error) {
@@ -54,7 +63,37 @@ exports.loginUser = async (req, res, next) => {
   const { username, password } = req.body;
   try {
     const user = await UserService.loginUser(username, password);
-    res.status(200).json({ ...new UserDto(user) });
+    const payload = { ...new UserDto(user) };
+    const { accessToken, refreshToken } = TokenService.createTokens(payload);
+    await TokenService.saveRefreshToken(payload, refreshToken);
+    res.cookie("refreshToken", refreshToken, {
+      maxAge: 30 * 24 * 60 * 60 * 1000,
+      httpOnly: true,
+    });
+    res.status(200).json({ accessToken });
+  } catch (error) {
+    next(error);
+  }
+};
+
+exports.logoutUser = async (req, res, next) => {
+  const { refreshToken } = req.cookies;
+  if (!refreshToken) {
+    return next(ApiError.Unauthorized("No logged-in user"));
+  }
+  await UserService.logoutUser(refreshToken);
+  res.clearCookie("refreshToken");
+  res.sendStatus(200);
+};
+
+exports.refresh = async (req, res, next) => {
+  const { refreshToken } = req.cookies;
+  try {
+    if (!refreshToken) {
+      return next(ApiError.Unauthorized("No logged-in user"));
+    }
+    const accessToken = await UserService.refresh(refreshToken);
+    res.status(200).json({ accessToken });
   } catch (error) {
     next(error);
   }
